refactor: KakaoMap 위치 변경, 현재 위치 button -> Button으로 수정
