+++ src/app/map/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import KakaoMap from '../components/KakaoMap';
+import KakaoMap from './KakaoMap';
 
 const Map = () => <KakaoMap />;
